Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 #include<stdio.h>
 int main()
 {
-    int a,b,c,avg;
+    float a,b,c,avg;
     scanf("%f %f %f",&a,&b,&c);
     avg=(a+b+c)/3;
     printf("Average: %f",avg);
